Add the FSID to the $cluster_option variable in osd.pp.
If you are redploying on top of hardware that has previously hosted
an OSD then it is possible for the old FSID to linger. If this happens
then a deployment may have no OSDs and when investigating you may see
the following error:

[root@overcloud-cephstorage-0 ~]# ceph-disk list
/dev/sda :
 /dev/sda1 other, iso9660
 /dev/sda2 other, xfs, mounted on /
/dev/sdb :
 /dev/sdb1 ceph data, prepared, unknown cluster 7756567c-97d6-11e5-8cf3-52540089f28a, osd.11, journal /dev/sdn4
/dev/sdc :
 /dev/sdc1 ceph data, prepared, unknown cluster 7756567c-97d6-11e5-8cf3-52540089f28a, osd.5, journal /dev/sdn1
/dev/sdd :
...
[root@overcloud-cephstorage-0 ~]#

This change explicitly adds the FSID to the $cluster_option variable,
if the FSID is defined. Else, leaves the $cluster_option variable as is.